==> 02_Analyses/02a_ER_all.Rmd <==
---
title: "ER Analyses"
author: "JMH"
date: "9 Dec 2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo= FALSE}
library(tidyverse)
library(here)
library(mgcViz)
library(MuMIn)
library(mgcv)
library(forecast)
library(gridExtra)
library(tidymv)
library(ggpubr)
```

# Load & munge data
```{r}
load(file.path(here::here("04_SavedImages"), "02_ERanalyses_rdat"))
```


```{r message=FALSE, warning=FALSE}
met <- readr::read_csv(file = file.path(here::here("01_Data"),"01_MundgedMetabDat.csv")) %>% 
  mutate(stream = fct_relevel(stream, c("st11U", "st18", "st6", "st9")),
         treatment = fct_relevel(treatment, c("ambient", "nitrogen", "phosphorus")),
         Yf = as.factor(Yf),
         streamTreat = as.factor(streamTreat),
         ppdoy180 = ifelse(doy <180,"l180", "g180"))%>%
  # cutting it off here removes most of the weird, low ER values
  filter(doy >=183) %>%
  group_by(streamTreat) %>% 
  dplyr::mutate(DD = cumsum(meanTemp)) %>% 
  ungroup() %>% 
  droplevels()

```



# ER Plots
Big differences among treatments. Normalish
```{r echo=FALSE}
ggplot(met, aes(y = lER, x = doy, color = treatment, shape = ppdoy180)) +
  geom_point()+
  stat_smooth() +
  facet_wrap(vars(stream))
```

```{ recho=FALSE}
ggplot(met, aes(y = meanQds, x = doy, color = treatment, shape = ppdoy180)) +
  geom_point()+
  stat_smooth() +
  facet_wrap(vars(stream))
```


```{r echo=FALSE}
ggplot(met, aes(sample = lER)) +stat_qq() + stat_qq_line() 
```


```{r echo=FALSE}
ggplot(met, aes(sample = lER, color = stream)) +stat_qq() + stat_qq_line() +
  facet_grid(.~stream)
```



Futzing to find a well-behaved global model
```{r include = FALSE}
# numerous models were evaluated in earlier versions
g.er.global <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
                       s(Qres, by = stream) + s(DOYres, by = stream),
                     REML = FALSE, data = met,
                     control = lmeControl(maxIter = 1e8, msMaxIter = 1e8))

```



```{r echo=FALSE}
summary(g.er.global$gam)
```

```{r echo=FALSE}
g.er.global.gv <- getViz(g.er.global$gam)
print(plot(g.er.global.gv, allTerms = T), pages =1)
```

```{r echo=FALSE}
check.gamViz(g.er.global.gv)
```


Autocorrelation - bad
```{r echo= FALSE}

# https://www.kaggle.com/timib1203/managing-many-models-with-tidyr-purrr-and-broom
# https://jroy042.github.io/nonlinear/week4.html
# https://stackoverflow.com/questions/47586110/autocorrelation-in-generalized-additive-models-gam
g.er.global.acf <- met %>% 
  mutate(g.erRes = resid(g.er.global$lme, type = "normalized")) %>% 
  select(streamTreat, g.erRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.erRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.er.global.acf$acf, ncol = 3)
```

Looks pretty good
Reassuring discussion of concurvity: https://m-clark.github.io/generalized-additive-models/issues.html
Longer discussion in Amodio et al. 2014
```{r echo=FALSE}
round(concurvity(g.er.global$gam, full = FALSE)$estimate,3)
```
Interesting thread on colinearity: https://stats.stackexchange.com/questions/100175/when-can-we-speak-of-collinearity/100272#100272

## Determine ARMA model
good explanation about why this is important in gams: https://fromthebottomoftheheap.net/2011/07/21/smoothing-temporally-correlated-data/

Preliminary analysis indicated that there was significant autocorrelation. So the first step is to determine the autocorrelation structure. We'll use what is sort of a global model.

```{r echo = FALSE, include = FALSE}
# g.er.AC_1.0 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                        s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 1,q = 0, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), REML = FALSE, data = met)
# 
# g.er.AC_1.1 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                        s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 1,q = 1, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), REML = FALSE, data = met)
# 
# g.er.AC_1.2 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
#                        s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 1,q = 2, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), REML = FALSE, data = met)
# 
# # g.er.AC_2.0 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
# #                       s(Qres, by = stream) + s(DOYres, by = stream),
# #                 correlation = corARMA(p = 2,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
# #                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_2.1 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 2,q = 1, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), REML = FALSE, data = met)
# 
# g.er.AC_2.2 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 2,q = 2, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), REML = FALSE, data = met)
# # 
# g.er.AC_3.0 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 3,q = 0, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), REML = FALSE, data = met)
# 
# g.er.AC_3.1 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 3,q = 1, form = ~ doy | streamTreat),
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), REML = FALSE, data = met)
# 
# g.er.AC_3.2 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 3,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_3.3 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 3,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_3.4 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 3,q = 4, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_4.0 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 4,q =0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# # g.er.AC_4.1 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
# #                       s(Qres, by = stream) + s(DOYres, by = stream),
# #                 correlation = corARMA(p = 4,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
# #                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_4.2 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 4,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_4.3 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 4,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# # g.er.AC_4.4 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
# #                       s(Qres, by = stream) + s(DOYres, by = stream),
# #                 correlation = corARMA(p = 4,q = 4, form = ~ doy | streamTreat), REML = FALSE, data = met,
# #                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))
# 
# g.er.AC_5.3 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 5,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.er.AC_6.3 <- gamm(lER ~ treatment + invKT.C.StMean + Tanom.iktCs +  
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = corARMA(p = 6,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500))

```

Model selection indicates that 3.4 is the best model.
```{r echo= FALSE}
model.sel(g.er.AC_1.0$lme, 
          g.er.AC_1.1$lme,
          g.er.AC_1.2$lme, 
          # g.er.AC_2.0$lme, 
          g.er.AC_2.1$lme, 
          g.er.AC_2.2$lme, 
          g.er.AC_3.0$lme,
          g.er.AC_3.1$lme, 
          g.er.AC_3.2$lme,
          g.er.AC_3.3$lme,
          g.er.AC_3.4$lme, 
          g.er.AC_4.0$lme,
          # g.er.AC_4.1$lme, 
          g.er.AC_4.2$lme, g.er.AC_4.3$lme, 
          # g.er.AC_4.4$lme, 
          g.er.AC_5.3$lme,
          g.er.AC_6.3$lme) 
```


ACF: Much better. S6P only one with minor problems
```{r echo= FALSE}

# https://www.kaggle.com/timib1203/managing-many-models-with-tidyr-purrr-and-broom
# https://jroy042.github.io/nonlinear/week4.html
# https://stackoverflow.com/questions/47586110/autocorrelation-in-generalized-additive-models-gam
g.er.AC_best.acf <- met %>% 
  mutate(g.erRes = resid(g.er.AC_3.3$lme, type = "normalized")) %>% 
  select(streamTreat, g.erRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.erRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.er.AC_best.acf$acf, ncol = 3)
```

```{r echo=FALSE}
g.er.AC_3.3.gv <- getViz(g.er.AC_3.3$gam)
check(g.er.AC_3.3.gv)
```

# NEW MODEL SELECTION

```{r echo= FALSE}
BestERCor <- corARMA(p = 3,q = 3, form = ~ doy | streamTreat)
```


## Model sel

```{r echo=FALSE, include = FALSE}
# full with stream as temp or category
# g.er.modsel1t <-  uGamm(lER ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), class = "gamm")
# 
# g.er.modsel1s <-  uGamm(lER ~ treatment * stream  * Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres, by = stream),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), class = "gamm")
# 
# # Qres only with stream as temp or category
# g.er.modsel2t <-  uGamm(lER ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                       s(Qres) + s(DOYres, by = stream),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), class = "gamm")
# 
# g.er.modsel2s <-  uGamm(lER ~ treatment * stream * Tanom.iktCs +
#                       s(Qres) + s(DOYres, by = stream),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), class = "gamm")
# 
# # Doy only with stream as temp or category
# g.er.modsel3t <-  uGamm(lER ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 500), class = "gamm")
# 
# g.er.modsel3s <-  uGamm(lER ~ treatment * stream * Tanom.iktCs +
#                       s(Qres, by = stream) + s(DOYres),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.er.Dredge_1t <-  dredge(global.model = g.er.modsel1t, evaluate = TRUE, rank = AIC)
# g.er.Dredge_1s <-  dredge(global.model = g.er.modsel1s, evaluate = TRUE, rank = AIC)
# g.er.Dredge_2t <-  dredge(global.model = g.er.modsel2t, evaluate = TRUE, rank = AIC)
# g.er.Dredge_2s <-  dredge(global.model = g.er.modsel2s, evaluate = TRUE, rank = AIC)
# g.er.Dredge_3t <-  dredge(global.model = g.er.modsel3t, evaluate = TRUE, rank = AIC)
# g.er.Dredge_3s <-  dredge(global.model = g.er.modsel3s, evaluate = TRUE, rank = AIC)
```



```{r echo=FALSE}
g.erDredgeSumt1 <- merge(g.er.Dredge_1t, g.er.Dredge_2t)
g.erDredgeSumtF <- merge(g.erDredgeSumt1, g.er.Dredge_3t)
g.erDredgeSums1 <- merge(g.er.Dredge_1s, g.er.Dredge_2s)
g.erDredgeSumsF <- merge(g.erDredgeSums1, g.er.Dredge_3s)
g.erDredgeSumF <- merge(g.erDredgeSumtF, g.erDredgeSumsF)
```

Model selection for all models
```{r echo=FALSE}
subset(g.erDredgeSumF, delta <= 2)
```

```{r}
g.er.BEST <- gamm(lER ~ treatment * stream + treatment * Tanom.iktCs + stream * Tanom.iktCs  +
                      s(Qres, by = stream) + s(DOYres, by = stream),
                correlation = BestERCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))


```

Stream Temperature models; AIC difference of -311.1 - -301.9 
```{r echo=FALSE}
subset(g.erDredgeSumtF, delta <= 2)
```


Best with stream temp
```{r}
g.er.BEST_t <- gamm(lER ~ treatment * Tanom.iktCs + invKT.C.StMean * Tanom.iktCs  +
                      s(Qres, by = stream) + s(DOYres),
                correlation = BestERCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```

### Best models
#### Overall
lER ~ treatment * stream + treatment * Tanom.iktCs + stream * Tanom.iktCs + 
      s(Qres, by = stream) + s(DOYres, by = stream)
R2 = 0.92
Sig? This is a pain to interpret! Not going to bother.

```{r}
summary(g.er.BEST$gam)
```


```{r}
round(concurvity(g.er.BEST$gam, full = FALSE)$estimate,3)
```

```{r}
g.er.BEST.gv <- getViz(g.er.BEST$gam)
```

```{r}
check(g.er.BEST.gv)
```

One last check on the autocorrelation. Just foucsing on 3. Some remaining
```{r echo= FALSE}
g.er.best.acf <- met %>% 
  mutate(g.erRes = resid(g.er.BEST$lme, type = "normalized")) %>% 
  select(streamTreat, g.erRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.erRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.er.best.acf$acf, ncol = 3)
```


```{r echo=FALSE}
metBestTest <- met %>% 
  mutate(erres1 = residuals(g.er.BEST$gam, type = "response"),
         erPred1 = predict(g.er.BEST$gam, type = "response"),
         erres2 = residuals(g.er.BEST_t$gam, type = "response"),
         erPred2 = predict(g.er.BEST_t$gam, type = "response"))
```


 Check residuals
Still some signal with doy. Also early dates with S18-P giving problems
Not much of a difference between 1 and 2.
```{r echo=FALSE}
ggplot(metBestTest, aes(y = erres1, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = lER, x = erPred1, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```


#### stream temp


lER ~ treatment * Tanom.iktCs + invKT.C.StMean * Tanom.iktCs + 
    s(Qres, by = stream) + s(DOYres)
R2 = 0.758
Sig effect of N and P
Sig daily and annual temp
Sig interaction between P and daily temp
```{r}
summary(g.er.BEST_t$gam)
```

```{r}
round(concurvity(g.er.BEST_t$gam, full = FALSE)$estimate,3)
```

```{r}
g.er.BEST_t.gv <- getViz(g.er.BEST_t$gam)
check(g.er.BEST_t.gv)
```

Still some autocorrelation remaining
```{r echo= FALSE}
g.er.BEST_t.acf <- met %>% 
  mutate(g.erRes = resid(g.er.BEST_t$lme, type = "normalized")) %>% 
  select(streamTreat, g.erRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.erRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.er.BEST_t.acf$acf, ncol = 3)
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = erres2, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```


```{r echo=FALSE}
ggplot(metBestTest, aes(y = lER, x = erPred2, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```



### save results
```{r echo = FALSE, include = FALSE}
saveRDS(g.er.BEST,
        file = file.path(here::here("05_SavedModels"), "02a_ERall_Best.rds"))
saveRDS(g.er.BEST_t,
        file = file.path(here::here("05_SavedModels"), "02a_ERall_BestTemp.rds"))
```


```{r echo = FALSE, include = FALSE}
# save.image(file.path(here::here("04_SavedImages"), "02_ERanalyses_rdat"))
```
