==> 02_Analyses/03a_GPP_all.Rmd <==
---
title: "GPP Analyses"
author: "JMH"
date: "11/19/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo= FALSE}
library(tidyverse)
library(here)
library(mgcViz)
library(MuMIn)
library(mgcv)
library(forecast)
library(gridExtra)
```

# Load & munge data
```{r}
load(file.path(here::here("04_SavedImages"), "03a_GPPanalyses_rdat"))
```


DOY < 183 already removed
```{r message=FALSE, warning=FALSE}
met <- readr::read_csv(file = file.path(here::here("01_Data"),"02_PostERmet.csv")) %>% 
  mutate(stream = fct_relevel(stream, c("st11U", "st18", "st6", "st9")),
         treatment = fct_relevel(treatment, c("ambient", "nitrogen", "phosphorus")),
         Yf = as.factor(Yf),
         streamTreat = as.factor(streamTreat))
```

# GPP Plots
Big differences among treatments. Pretty normal
```{r echo=FALSE}
ggplot(met, aes(y = lGPP, x = doy, color = treatment)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(vars(stream))
```

```{r echo=FALSE}
ggplot(met, aes(sample = lGPP, color = stream)) +stat_qq() + stat_qq_line() +
  facet_grid(.~stream)
```

# GAMMS

## Determine ARMA model

Preliminary analysis indicated that there was significant autocorrelation. So the first step is to determine the autocorrelation structure. We'll use what is sort of a global model.

```{r}
# g.gpp.AC_1.0 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 1,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_1.1 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 1,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_1.2 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 1,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_2.0 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 2,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_2.1 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 2,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_2.2 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 2,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_3.0 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_3.1 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_3.2 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_3.3 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_3.4 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg),
#                 correlation = corARMA(p = 3,q = 4, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8), niterEM = 1000)
# 
# g.gpp.AC_4.0 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 4,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# # g.gpp.AC_4.1 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
# #                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
# #                 correlation = corARMA(p = 4,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
# #                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# # g.gpp.AC_4.2 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
# #                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
# #                 correlation = corARMA(p = 4,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
# #                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_4.3 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 4,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.gpp.AC_4.4 <- gamm(lGPP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 4,q = 4, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))

```

Model selection indicates that 3_2 is the best model.
```{r echo= FALSE}
model.sel(g.gpp.AC_1.0$lme, g.gpp.AC_1.1$lme, g.gpp.AC_1.2$lme, 
          g.gpp.AC_2.0$lme, g.gpp.AC_2.1$lme, g.gpp.AC_2.2$lme, 
          g.gpp.AC_3.0$lme, g.gpp.AC_3.1$lme, g.gpp.AC_3.2$lme, g.gpp.AC_3.3$lme,
          g.gpp.AC_4.0$lme, 
          # g.gpp.AC_4.1$lme,g.gpp.AC_4.2$lme, 
          g.gpp.AC_4.3$lme) 
```

ACF: Looks great. One sig lag in S9P and S18N
```{r echo= FALSE}

# https://www.kaggle.com/timib1203/managing-many-models-with-tidyr-purrr-and-broom
# https://jroy042.github.io/nonlinear/week4.html
# https://stackoverflow.com/questions/47586110/autocorrelation-in-generalized-additive-models-gam
g.gpp.AC_best.acf <- met %>% 
  mutate(g.gppRes = resid(g.gpp.AC_3.2$lme, type = "normalized")) %>% 
  select(streamTreat, g.gppRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.gppRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.gpp.AC_best.acf$acf, ncol = 3)
```

```{r echo= FALSE}
BestGPPCor <- corARMA(p = 3,q = 2, form = ~ doy | streamTreat)
```



## Model selection
```{r echo=FALSE, include = FALSE}
# g.gpp.modsel1t <- uGamm(lGPP ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestGPPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.gpp.modsel1s <- uGamm(lGPP ~ treatment * stream * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestGPPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# # full interaction wouldn't converge
# # and neither would the 3 interaction model
# g.gpp.modsel2t <- uGamm(lGPP ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                      s(Qres) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestGPPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.gpp.modsel2s <- uGamm(lGPP ~ treatment * stream * Tanom.iktCs +
#                      s(Qres) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestGPPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.gpp.modsel3t <- uGamm(lGPP ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres) + s(LightPerDay.lcg),
#                 correlation = BestGPPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.gpp.modsel3s <- uGamm(lGPP ~ treatment * stream * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres) + s(LightPerDay.lcg),
#                 correlation = BestGPPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.gpp.Dredge_1t <-  dredge(global.model = g.gpp.modsel1t, evaluate = TRUE, rank = AIC)
# g.gpp.Dredge_1s <-  dredge(global.model = g.gpp.modsel1s, evaluate = TRUE, rank = AIC)
# g.gpp.Dredge_2t <-  dredge(global.model = g.gpp.modsel2t, evaluate = TRUE, rank = AIC)
# g.gpp.Dredge_2s <-  dredge(global.model = g.gpp.modsel2s, evaluate = TRUE, rank = AIC)
# g.gpp.Dredge_3t <-  dredge(global.model = g.gpp.modsel3t, evaluate = TRUE, rank = AIC)
# g.gpp.Dredge_3s <-  dredge(global.model = g.gpp.modsel3s, evaluate = TRUE, rank = AIC)
```


```{r echo=FALSE}
g.gppDredgeSumT1 <- merge(g.gpp.Dredge_1t, g.gpp.Dredge_2t)
g.gppDredgeSumT2 <- merge(g.gppDredgeSumT1, g.gpp.Dredge_3t)
g.gppDredgeSumS1 <- merge(g.gpp.Dredge_1s, g.gpp.Dredge_2s)
g.gppDredgeSumS2 <- merge(g.gppDredgeSumS1, g.gpp.Dredge_3s)
g.gppDredgeSumF <- merge(g.gppDredgeSumT2, g.gppDredgeSumS2)
```


Model selection for all models
```{r echo=FALSE}
subset(g.gppDredgeSumF, delta <= 2)
```

```{r}
g.gpp.BEST <- gamm(lGPP ~ treatment + stream * Tanom.iktCs +
                     s(Qres, by = stream) + s(DOYres) + s(LightPerDay.lcg),
                correlation = BestGPPCor, REML = FALSE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```

Stream Temperature models; AIC difference 2.2
```{r echo=FALSE}
subset(g.gppDredgeSumT2, delta <= 2)
```

Best with stream temp
```{r}
g.gpp.BEST_t <- gamm(lGPP ~ treatment + invKT.C.StMean * Tanom.iktCs +
                     s(Qres, by = stream) + s(DOYres) + s(LightPerDay.lcg),
                correlation = BestGPPCor, REML = FALSE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```

### Best models
#### Overall
 
lGPP ~ treatment + stream * Tanom.iktCs + s(Qres, by = stream) + 
    s(DOYres) + s(LightPerDay.lcg)
R2 = 0.69
N effect and P effect. No interaction with stream or daily T
GPP decreases with daily T in S6

```{r}
summary(g.gpp.BEST$gam)
```

Not bad
```{r}
round(concurvity(g.gpp.BEST$gam, full = FALSE)$estimate,3)
```

```{r}
g.gpp.BEST.gv <- getViz(g.gpp.BEST$gam)
check(g.gpp.BEST.gv)
```



One last check on the autocorrelation. Just foucsing on 3. Some remaining in S9A, S9P, S11UA, S18N

```{r echo= FALSE, include = FALSE}
g.gpp.best.acf <- met %>%
  mutate(g.gppRes = resid(g.gpp.BEST$lme, type = "normalized")) %>%
  select(streamTreat, g.gppRes) %>%
  group_by(streamTreat) %>%
  nest() %>%
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.gppRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.gpp.best.acf$acf, ncol = 3)
```


```{r echo=FALSE}
metBestTest <- met %>% 
  mutate(gppres1 = residuals(g.gpp.BEST$gam, type = "response"),
         gppPred1 = predict(g.gpp.BEST$gam, type = "response"),
         gppres2 = residuals(g.gpp.BEST_t$gam, type = "response"),
         gppPred2 = predict(g.gpp.BEST_t$gam, type = "response"))
```


 Check residuals
Still some signal with doy. Also early dates with S18-P giving problems
Not much of a difference between 1 and 2.
```{r echo=FALSE}
ggplot(metBestTest, aes(y = gppres1, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = lGPP, x = gppPred1, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```


#### stream temp


lGPP ~ treatment + invKT.C.StMean * Tanom.iktCs + s(Qres, by = stream) + 
    s(DOYres) + s(LightPerDay.lcg)
R2 = 0.616
Sig effect of N 
Sig effect of stream T and daily T, interaction ?decreases? GPP
```{r}
summary(g.gpp.BEST_t$gam)
```

looks pretty good
```{r}
round(concurvity(g.gpp.BEST_t$gam, full = FALSE)$estimate,3)
```

```{r}
g.gpp.BEST_t.gv <- getViz(g.gpp.BEST_t$gam)
check(g.gpp.BEST_t.gv)
```

Still some autocorrelation remaining in S9A, A9P, S18N
```{r echo= FALSE}
g.gpp.BEST_t.acf <- met %>% 
  mutate(g.gppRes = resid(g.gpp.BEST_t$lme, type = "normalized")) %>% 
  select(streamTreat, g.gppRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.gppRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.gpp.BEST_t.acf$acf, ncol = 3)
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = gppres2, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```


```{r echo=FALSE}
ggplot(metBestTest, aes(y = lGPP, x = gppPred2, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```



### save results
```{r echo = FALSE, include = FALSE}
 # saveRDS(g.gpp.BEST,
 #         file = file.path(here::here("05_SavedModels"), "03a_GPPall_Best.rds"))
 # saveRDS(g.gpp.BEST_t,
 #         file = file.path(here::here("05_SavedModels"), "03a_GPPall_BestTemp.rds"))
```


```{r}
# save.image(file.path(here::here("04_SavedImages"), "03a_GPPanalyses_rdat"))
```
