==> 02_Analyses/04a_NEP_all.Rmd <==
---
title: "NEP Analyses"
author: "JMH"
date: "11/19/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo= FALSE}
library(tidyverse)
library(here)
library(mgcViz)
library(MuMIn)
library(mgcv)
library(forecast)
library(gridExtra)
```

# Load & munge data
```{r}
load(file.path(here::here("04_SavedImages"), "04a_NEP_all_rdat"))
```


DOY < 183 already removed
```{r message=FALSE, warning=FALSE}
met <- readr::read_csv(file = file.path(here::here("01_Data"),"02_PostERmet.csv")) %>% 
  mutate(stream = fct_relevel(stream, c("st11U", "st18", "st6", "st9")),
         treatment = fct_relevel(treatment, c("ambient", "nitrogen", "phosphorus")),
         Yf = as.factor(Yf),
         streamTreat = as.factor(streamTreat))
```

# NEP Plots
Big differences among treatments. Pretty normal
```{r echo=FALSE}
ggplot(met, aes(y = NEP, x = doy, color = treatment)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(vars(stream))
```

```{r echo=FALSE}
ggplot(met, aes(sample = NEP, color = stream)) +stat_qq() + stat_qq_line() +
  facet_grid(.~stream)
```

# GAMMS

## Determine ARMA model

Preliminary analysis indicated that there was significant autocorrelation. So the first step is to determine the autocorrelation structure. We'll use what is sort of a global model.

```{r}
# g.NEP.AC_1.0 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 1,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_1.1 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 1,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_1.2 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 1,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_2.0 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 2,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_2.1 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 2,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_2.2 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 2,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_3.0 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_3.1 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_3.2 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_3.3 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 3,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_3.4 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg),
#                 correlation = corARMA(p = 3,q = 4, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8), niterEM = 1000)
# 
# g.NEP.AC_4.0 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 4,q = 0, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
#  g.NEP.AC_4.1 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg),
#                  correlation = corARMA(p = 4,q = 1, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                  control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
#  g.NEP.AC_4.2 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +
#                       s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg),
#                  correlation = corARMA(p = 4,q = 2, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                  control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_4.3 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 4,q = 3, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
# 
# g.NEP.AC_4.4 <- gamm(NEP ~ treatment + invKT.C.StMean + Tanom.iktCs +   
#                      s(Qres, by = stream) + s(doy, by = stream) + s(LightPerDay.lcg), 
#                 correlation = corARMA(p = 4,q = 4, form = ~ doy | streamTreat), REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))

```

Model selection indicates that 4_2 is the best model.
```{r echo= FALSE}
model.sel(g.NEP.AC_1.0$lme, g.NEP.AC_1.1$lme, g.NEP.AC_1.2$lme, 
          g.NEP.AC_2.0$lme, g.NEP.AC_2.1$lme, g.NEP.AC_2.2$lme, 
          g.NEP.AC_3.0$lme, g.NEP.AC_3.1$lme, g.NEP.AC_3.2$lme, g.NEP.AC_3.3$lme,
          g.NEP.AC_4.0$lme, 
          g.NEP.AC_4.1$lme,g.NEP.AC_4.2$lme,
          g.NEP.AC_4.3$lme) 
```

ACF: Looks great. One sig lag in S9A, S18N, S18P
```{r echo= FALSE}

# https://www.kaggle.com/timib1203/managing-many-models-with-tidyr-purrr-and-broom
# https://jroy042.github.io/nonlinear/week4.html
# https://stackoverflow.com/questions/47586110/autocorrelation-in-generalized-additive-models-gam
g.NEP.AC_best.acf <- met %>% 
  mutate(g.gppRes = resid(g.NEP.AC_4.2$lme, type = "normalized")) %>% 
  select(streamTreat, g.gppRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.gppRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.NEP.AC_best.acf$acf, ncol = 3)
```

```{r echo= FALSE}
BestNEPCor <- corARMA(p = 4,q = 2, form = ~ doy | streamTreat)
```



## Model selection
```{r echo=FALSE, include = FALSE}
# g.NEP.modsel1t <- uGamm(NEP ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestNEPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.NEP.modsel1s <- uGamm(NEP ~ treatment * stream * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestNEPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# 
# g.NEP.modsel2t <- uGamm(NEP ~ treatment * invKT.C.StMean * Tanom.iktCs +
#                      s(Qres) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestNEPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# # these models won't converge, so tried one without full interaction
# g.NEP.modsel2s <- uGamm(NEP ~ treatment * stream + stream * Tanom.iktCs + treatment * Tanom.iktCs +
#                      s(Qres) + s(DOYres, by = stream) + s(LightPerDay.lcg),
#                 correlation = BestNEPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# # these models won't converge, so tried one without full interaction
# g.NEP.modsel3t <- uGamm(NEP ~ treatment * stream + stream * Tanom.iktCs + treatment * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres) + s(LightPerDay.lcg),
#                 correlation = BestNEPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# # these models won't converge, so tried one without full interaction
# g.NEP.modsel3s <- uGamm(NEP ~ treatment * stream + stream * Tanom.iktCs + treatment * Tanom.iktCs +
#                      s(Qres, by = stream) + s(DOYres) + s(LightPerDay.lcg),
#                 correlation = BestNEPCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.NEP.Dredge_1t <-  dredge(global.model = g.NEP.modsel1t, evaluate = TRUE, rank = AIC)
# g.NEP.Dredge_1s <-  dredge(global.model = g.NEP.modsel1s, evaluate = TRUE, rank = AIC)
# g.NEP.Dredge_2t <-  dredge(global.model = g.NEP.modsel2t, evaluate = TRUE, rank = AIC)
# g.NEP.Dredge_2s <-  dredge(global.model = g.NEP.modsel2s, evaluate = TRUE, rank = AIC)
# g.NEP.Dredge_3t <-  dredge(global.model = g.NEP.modsel3t, evaluate = TRUE, rank = AIC)
# g.NEP.Dredge_3s <-  dredge(global.model = g.NEP.modsel3s, evaluate = TRUE, rank = AIC)
```


```{r echo=FALSE}
g.NEPDredgeSumT1 <- merge(g.NEP.Dredge_1t, g.NEP.Dredge_2t)
g.NEPDredgeSumT2 <- merge(g.NEPDredgeSumT1, g.NEP.Dredge_3t)
g.NEPDredgeSumS1 <- merge(g.NEP.Dredge_1s, g.NEP.Dredge_2s)
g.NEPDredgeSumS2 <- merge(g.NEPDredgeSumS1, g.NEP.Dredge_3s)
g.NEPDredgeSumF <- merge(g.NEPDredgeSumT2, g.NEPDredgeSumS2)
```



Model selection for all models
```{r echo=FALSE}
subset(g.NEPDredgeSumF, delta <= 2)
```

```{r}
g.nep.BEST <- gamm(NEP ~ treatment * stream * Tanom.iktCs + 
                     s(Qres, by = stream) + s(DOYres, by = stream) + s(LightPerDay.lcg),
                correlation = BestNEPCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```

Stream Temperature models; this is second best 1.91
```{r echo=FALSE}
subset(g.NEPDredgeSumT2, delta <= 2)
```

2nd best with stream temp. 
```{r}
g.nep.BEST_t <- gamm(NEP ~ treatment * invKT.C.StMean * Tanom.iktCs +
                     s(Qres, by = stream) + s(DOYres, by = stream) + s(LightPerDay.lcg),
                correlation = BestNEPCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```

### Best models
#### Overall
 
NEP ~ treatment * stream * Tanom.iktCs + s(Qres, by = stream) + 
    s(DOYres, by = stream) + s(LightPerDay.lcg)

R2 = 0.82
What a nightmare
This is tough to explain since sig stream effect means its differnt from S11U
Sig N marginal P

```{r}
summary(g.nep.BEST$gam)
```

Issues with Qres and DOYres
```{r}
round(concurvity(g.nep.BEST$gam, full = FALSE)$estimate,3)
```

```{r}
g.nep.BEST.gv <- getViz(g.nep.BEST$gam)
check(g.nep.BEST.gv)
```



One last check on the autocorrelation. Just foucsing on 3. Some remaining in S11U P

```{r echo= FALSE, include = FALSE}
g.nep.best.acf <- met %>%
  mutate(g.nepRes = resid(g.nep.BEST$lme, type = "normalized")) %>%
  select(streamTreat, g.nepRes) %>%
  group_by(streamTreat) %>%
  nest() %>%
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.nepRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.nep.best.acf$acf, ncol = 3)
```


```{r echo=FALSE}
metBestTest <- met %>% 
  mutate(nepres1 = residuals(g.nep.BEST$gam, type = "response"),
         nepPred1 = predict(g.nep.BEST$gam, type = "response"),
         nepres2 = residuals(g.nep.BEST_t$gam, type = "response"),
         nepPred2 = predict(g.nep.BEST_t$gam, type = "response"))
```


 Check residuals
Still some signal with doy. Also early dates with S18-P giving problems
Not much of a difference between 1 and 2.
```{r echo=FALSE}
ggplot(metBestTest, aes(y = nepres1, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = NEP, x = nepPred1, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```


#### stream temp


NEP ~ treatment * invKT.C.StMean * Tanom.iktCs + s(Qres, by = stream) + 
    s(DOYres, by = stream) + s(LightPerDay.lcg)
R2 = 0.65
N x stream T x daily T SIG
No sig interactions with P, although P alone is sig

```{r}
summary(g.nep.BEST_t$gam)
```

Issues with doy and qres
```{r}
round(concurvity(g.nep.BEST_t$gam, full = FALSE)$estimate,3)
```


Some uglyness here
```{r}
g.nep.BEST_t.gv <- getViz(g.nep.BEST_t$gam)
check(g.nep.BEST_t.gv)
```

Still some autocorrelation remaining in S18N
```{r echo= FALSE}
g.nep.BEST_t.acf <- met %>% 
  mutate(g.nepRes = resid(g.nep.BEST_t$lme, type = "normalized")) %>% 
  select(streamTreat, g.nepRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.nepRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.nep.BEST_t.acf$acf, ncol = 3)
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = nepres2, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```


```{r echo=FALSE}
ggplot(metBestTest, aes(y = NEP, x = nepPred2, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```




### save results
```{r echo = FALSE, include = FALSE}
 saveRDS(g.nep.BEST,
         file = file.path(here::here("05_SavedModels"), "03a_NEP_all_Best.rds"))
 saveRDS(g.nep.BEST_t,
         file = file.path(here::here("05_SavedModels"), "03a_NEP_all_BestTemp.rds"))
```




```{r}
save.image(file.path(here::here("04_SavedImages"), "04a_NEP_all_rdat"))
# load(file.path(here::here("04_SavedImages"), "04a_NEP_all_rdat"))
```
